update: updated todo.

+++ src/modules/screen.c
@@ -92,7 +92,7 @@ int rgb_frame_brightness(const uint8_t *data, const int width, const int height,
     const int div = 8;
     const int wmax = (double)width / div;
     const int hmax = (double)height / div;
-    const int pixelsize = stride / width;
+    const int pixelsize = (double)stride / width;
     for (int i = 0; i < wmax; i++) {
         for (int k = 0; k < hmax; k++) {
             /* obtain r,g,b components */
